fix: deduplicate FORTRAN 66/77 statement function tests (fixes #221)
Add StatementFunctionTestMixin to shared_fortran_tests.py that
encapsulates common test patterns for statement function parsing
and array element disambiguation. Both FORTRAN66 and FORTRAN77
test classes now inherit from this mixin.

Also fixes test_proper_inheritance path resolution to use ROOT.parent
instead of relative path.
